backend/requirements.txt
# Core Framework
fastapi>=0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
fastapi-limiter>=0.1.5

# Database
psycopg2-binary>=2.9.9,<3.0.0
pymongo>=4.6.0,<5.0.0
motor>=3.3.0,<4.0.0  # Async MongoDB driver
redis>=5.0.1,<6.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# MCP (Model Context Protocol)
mcp==1.10.1

# AI & Machine Learning
langchain>=0.1.0,<0.2.0
langchain-mistralai

# Vector Search & Embeddings
sentence-transformers>=2.2.2,<3.0.0
faiss-cpu>=1.7.4,<2.0.0

# Utilities
httpx>=0.27.0
aiofiles>=23.2.1,<24.0.0
python-dateutil>=2.8.2,<3.0.0

# Development & Testing
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<0.22.0
black>=23.11.0,<24.0.0
flake8>=6.1.0,<7.0.0

# Monitoring & Logging
structlog>=23.2.0,<24.0.0
prometheus-client>=0.19.0,<0.20.0

# Production
gunicorn>=21.2.0,<22.0.0 
langgraph